Hub/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
paho-mqtt==1.6.1
pydantic==2.6.4
pydantic_core==2.16.3
redis==5.0.3
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
